+++ libs/shared/entity/src/lib/entity/language/language.enum.ts
@@ -1,6 +1,10 @@
 export enum Language {
-    English = 'en',
-    Dutch = 'nl',
-    French = 'fr',
-    Portuguese = 'pt',
-  }
+  English = 'en',
+  Dutch = 'nl',
+  French = 'fr',
+  Portuguese = 'pt',
+  Spanish = 'es',
+  German = 'de',
+  Arabic = 'ar',
+  Polish = 'pl',
+}