+++ libs/shared/translation/src/index.ts
@@ -3,5 +3,3 @@ export * from './lib/core/document-language';
 export * from './lib/core/translated-title.strategy';
 export * from './lib/core/language-storage';
 export * from './lib/core/update-language.service';
-export * from './lib/pages/update-language/update-language.component';
-export { updateLanguageRoutes } from './lib/pages/update-language/update-language-routes';